requirements.txt
absl-py==1.4.0
asn1crypto==0.24.0
cached-property==1.5.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.12
cryptography==2.1.4
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
dlib==19.24.2
face-alignment==1.2.0
facenet-pytorch==2.5.3
future==0.18.3
h5py==3.1.0
idna==2.6
imagecolor==1.2.1
imageio==2.15.0
imageio-ffmpeg==0.4.8
importlib-resources==5.4.0
joblib==1.1.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
lpips==0.1.4
matplotlib==3.3.4
mediapipe==0.8.2
mypy-extensions==1.0.0
networkx==2.5.1
ninja==1.11.1
numpy==1.19.3
nvdiffrast==0.3.1
opencv-python==4.7.0.72
pickle5==0.0.12
Pillow==8.4.0
pip==21.3.1
protobuf==3.19.6
pycrypto==2.6.1
PyGObject==3.26.1
pyparsing==3.0.9
pyrallis==0.3.1
python-apt==1.6.6
python-dateutil==2.8.2
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==6.0
redner-gpu==0.4.28
requests==2.27.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
SecretStorage==2.3.1
setuptools==39.0.1
sewar==0.4.6
six==1.11.0
sklearn==0.0.post7
ssh-import-id==5.7
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.0+cu111
torchaudio==0.10.0+rocm4.1
torchvision==0.11.0+cu111
tqdm==4.64.1
typing_extensions==4.1.1
typing-inspect==0.9.0
unattended-upgrades==0.1
urllib3==1.26.16
wheel==0.30.0
zipp==3.6.0
